Further tweaks/polishing
Updates to create the commit in a temp file (using os.tempDir()) and

then open that file; upon running the commit operation, we save the file

before closing the editor tab so the user doesn't get pestered about it.

Also updated to move the cursor to the beginning of the first line when

opening the commit message screen, for convenience.